refactor(styles): move to styles package
